% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{allele_distribution}
\alias{allele_distribution}
\title{Geographic distribution of alleles}
\usage{
allele_distribution(locus = "top-2", variant = 797, location = NA,
  colors = c("purple", "salmon"), alpha = 0.8)
}
\arguments{
\item{locus}{A gene or position. Examples: \code{top-2}, \code{II:14425175}, or "WBGene00010785"}

\item{variant}{character value corresponding amino acid change you are interested in plotting. Currently only supports visualization of SNVs that alter an amino acid residue.}

\item{location}{region to restrict plot to. Uses google api for lookup. Examples: \code{europe}.}

\item{colors}{character vector containing two colors. The first element of the vector will be the ALT genotype and the second will be the REF genotype. Default values are c("purple","salmon").}

\item{alpha}{Set marker alpha.}
}
\value{
Outputs a ggplot object containing a visualization of the global distribution of allele of interest
}
\description{
\code{allele_distribution} uses the ggplot2 package to plot the global distribution of an allele of interest.
}

