% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_mappings.R
\name{calculate_VE}
\alias{calculate_VE}
\title{Calcule Variance Explained for Significant SNPs}
\usage{
calculate_VE(mapping_df, snp_df = snps, phenotype_df)
}
\arguments{
\item{mapping_df}{the output from the \code{gwas_mappings} function. User input}

\item{snp_df}{genotype information for 124 wild isolates. Provided by package}

\item{phenotype_df}{two element list. element 1 : traits. element 2: trait values with strains in columns
with each row corresponding to trait in element 1}
}
\value{
Outputs a two element list that contains two dataframes.
The first data frame is a processed mappings dataframe that contains the same columns
as the output of \code{gwas_mappings} with two additional columns. One that contains
the bonferroni corrected p-value (BF) and another that contains an identifier 1,0 if
the indicated SNP has a higher -log10 value than the bonferroni cut off or not, respectively
The second data frame contains the variance explained data as well as all of the information from the first element.
}
\description{
\code{calculate_VE} calculates the variance explained (VE) for significant SNPs by using the
the spearman rank correlation coefficient.
}
\details{
This function requires three inputs, two of which are provided by the user and the other is loaded by the package.
}

