% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_correlation.R
\name{fetch_id_type}
\alias{fetch_id_type}
\title{Fetch Variant Type}
\usage{
fetch_id_type(id_type = NA)
}
\value{
Vector of identifiers.
}
\description{
\code{fetch_id_type} Fetches wormbase identifiers of a certain class. These can be piped into \code{snpeff}
}
\examples{
fetch_id_type("lincRNA")
}

