% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fine_map_plot}
\alias{fine_map_plot}
\title{Fine mapping plot of GWAS confidence interval}
\usage{
fine_map_plot(df, plot_trait, strain_comparison = "CB4856", color1 = "blue",
  color2 = "orange")
}
\arguments{
\item{df}{is a dataframe that is output by \code{process_correlations} function}

\item{plot_trait}{is a character vector corresponding to the trait of interest.}

\item{strain_comparison}{is a character vector corresponding to which strain genotype to use to color points by. Default is CB4856, which will color all variants that CB4856 contains a different color}

\item{color1}{is a character vector corresponding to the color of the ALT genotype points. Default is blue}

\item{color2}{is a character vector corresponding to the color of the REF genotype points. Default is orange}
}
\value{
returns a ggplot2 object
}
\description{
\code{fine_map_plot} Fine mapping plot of GWAS confidence interval
}
