% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gene_ids}
\alias{gene_ids}
\title{Gene IDs (Dataset)}
\format{list}
\usage{
gene_ids
}
\description{
A list of wormbase to standard gene identifier (e.g. _pot-2_) identifiers.
}
\keyword{datasets}

