% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcf.R
\name{generate_kinship}
\alias{generate_kinship}
\title{Generate Kinship}
\usage{
generate_kinship(vcf)
}
\arguments{
\item{vcf}{- a bcf/vcf/vcf.gz file}
}
\value{
A kinship matrix.
}
\description{
\code{generate_kinship} uses the rrBLUP package function \code{A.mat} to generate a kinship matrix for mapping.
Requires bcftools. Heterozygous calls and missing genotypes are ignored.
}

