% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{global_allele_distribution}
\alias{global_allele_distribution}
\title{Global distribution of allele}
\usage{
global_allele_distribution(gene = "top-2", variant = "797",
  colors = c("purple", "salmon"))
}
\arguments{
\item{gene}{character value gene of interest. Can be gene name "top-2" or gene id "WBGene00010785"}

\item{variant}{character value corresponding amino acid change you are interested in plotting. Currently only supports visualization of SNVs that alter an amino acid residue.}

\item{colors}{character vector containing two colors. The first element of the vector will be the ALT genotype and the second will be the REF genotype. Default values are c("purple","salmon").}
}
\value{
Outputs a ggplot object containing a visualization of the global distribution of allele of interest
}
\description{
\code{global_allele_distribution} uses the ggplot2 and ggmap packages to plot the global distribution of an allele of interest
}
\details{
This is the detail section if you want to fill out in the future
}

