% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mappings.R
\name{identify_CI}
\alias{identify_CI}
\title{Identify confidence intervals associated with QTL.}
\usage{
identify_CI(processed_mapping_df, peak_df, peak_list, correlation_df)
}
\arguments{
\item{processed_mapping_df}{The first element of the \code{calculate_VE} function output}

\item{peak_df}{The first element of the \code{find_peaks} function output}

\item{peak_list}{The second element of the \code{find_peaks} function output}

\item{correlation_df}{The second element of the \code{calculate_VE} function output}
}
\value{
Outputs processed mapping dataframe that contains original mapping dataframe with appended information for significant SNPs only, including:
variance explained, confidence interval information, genotype information
}
\description{
\code{identify_CI} Identifies confidence intervals for identified QTL
}
\details{
Function to combine all of the previously generated data into one data frame. Converts SNP index confidence intervals
into genomic position confidence intervals.
}

