% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{kinship}
\alias{kinship}
\title{Relatedness matrix (Dataset)}
\format{A Strain x Strain data frame for 124 strains}
\description{
Kinship matrix generated using the \code{a.mat} function in the rrBLUP package.
Whole-genome SNP data was used to generate the relatedness of the strains in this file.
}

