% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mapping_snps}
\alias{mapping_snps}
\title{Mapping SNPs (Dataset)}
\format{A data frame with snps as rows and columns as strains:
\describe{
  \item{CHROM}{chromosome in roman numerals}
  \item{POS}{physical position in genome}
  \item{AB1}{first strain in collection, followed by all other strains}
  ...
}}
\usage{
mapping_snps
}
\description{
Reduced SNP set
}
\keyword{datasets}

