% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_peak_ld}
\alias{plot_peak_ld}
\title{Plot LD values for significant SNPs in a mapping}
\usage{
plot_peak_ld(plot_df, trait = NULL)
}
\arguments{
\item{plot_df}{is the output from the process_mappings function.}

\item{trait}{is a string object corresponding to a trait of interest if plot_df has multiple traits in it.}
}
\value{
returns an LDHeatmap object of SNPs that are above the Bonferroni corrected p-value
}
\description{
\code{plot_peak_ld} generates a plot to visualize LD (r) for significant SNPs in a mapping
}
\examples{
 plot_peak_ld(processed_mapping_df) # for a one trait mapping data frame
}

