% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_correlation.R
\name{process_correlations}
\alias{process_correlations}
\title{Combine Variant Correlation Data}
\usage{
process_correlations(df)
}
\arguments{
\item{df}{is a list object that is output from the \code{variant_correlation} function}
}
\value{
Outputs a data frame that contains phenotype data, mapping data, and gene information for highly correlated variants in a particular QTL confidence interval.
}
\description{
\code{process_correlations} Combines all variant correlation information with phenotype data from GWAS mappings
}
\details{
Function to combine phenotype, mapping, and gene information from \code{variant_correlation} function output.
}

