% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_mappings.R
\name{process_mappings}
\alias{process_mappings}
\title{Fully process GWAS mapping output}
\usage{
process_mappings(mapping_df, phenotype_df, CI_size = 50, snp_grouping = 200,
  BF = NA)
}
\arguments{
\item{mapping_df}{Output from \code{gwas_mappings} function}

\item{phenotype_df}{phenotype data frame generated by \code{process_pheno}. two element list. element 1 : traits. element 2: trait values with strains in columns with each row corresponding to trait in element 1}

\item{CI_size}{defines the size (in # SNPs) of confidence intervals. Default is 50 and is defined in more detail below.}

\item{snp_grouping}{defines grouping of peaks. Defined further below, default is 200.}

\item{BF}{defines a custom bonferroni correction.}
}
\value{
Outputs processed mapping dataframe that contains original mapping dataframe with appended information for significant SNPs only, including: variance explained, confidence interval information, genotype information
}
\description{
\code{process_mappings} takes \code{gwas_mappings} output and calculates variance explained and
identifies peaks and associated confidence intervals
}
\details{
This function combines \code{calculate_VE}, \code{find_peaks}, and \code{identify_CI} into one function when intermediate dataframes are not wanted
}

