% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_phenotypes.R
\name{process_pheno}
\alias{process_pheno}
\title{Process Phenotype Data}
\usage{
process_pheno(data)
}
\arguments{
\item{data}{is a dataframe containing phenotype data. The first column should be names \code{trait}
all additional columns should be strains. One row corresponding to one trait for all strains.}
}
\value{
Outputs a list. The first element of the list is an ordered vector of traits.
The second element of the list is a dataframe containing one column for each strain, with values corresponding to traits in element 1 for rows.
}
\description{
\code{process_pheno} Process raw data file for GWAS mapping using \code{gwas_mappings} function
}
\details{
This function takes raw phenotype data and eliminates outlier strains with a modified version of \code{bamf_prune} from the easysorter package.
Additionally it eliminates any traits that have the same values for >95% of the strains (important for binary traits)
}

