% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{pxg_plot}
\alias{pxg_plot}
\title{PxG plot}
\usage{
pxg_plot(plot_df)
}
\arguments{
\item{plot_df}{the output from the \code{gwas_mappings} function.}
}
\value{
Ouput is a ggplot object facetted by peak SNP (if there are multiple peaks in a mapping).
Genotypes are encoded as 1 or 0 and are on the x-axis. Phenotypes are on y-axis.
}
\description{
\code{pxg_plot} generates a boxplot of phenotypes split by genotype at QTL peak position using \code{ggplot2}
}

