% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ-plot implemented in ggplot2}
\usage{
qq_plot(log10p)
}
\arguments{
\item{log10p}{vector of log10(p) values}
}
\value{
returns a ggplot2 object
}
\description{
\code{qq_plot} generates a QQ plot given a vector of log10(p) values
}

