% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{query_vcf}
\alias{query_vcf}
\title{Query VCF Data}
\usage{
query_vcf(..., info = c(), format = c("TGT"), impact = c("MODERATE",
  "HIGH"), samples = "ALL", long = TRUE, vcf = get_vcf())
}
\arguments{
\item{...}{Gene names, regions, or wormbase identifiers to query.}

\item{info}{Info columns to output. If an \code{ANN} (annotation) column is available it is automatically fetched. [\strong{Default} \code{c()}]}

\item{samples}{A set of samples to subset on [\strong{default:} \code{"ALL"}]}

\item{long}{Return dataset in long or wide format. [\strong{Default} \code{TRUE}]}

\item{vcf}{Use a custom VCF.}
}
\value{
Dataframe with variant data
}
\description{
\code{query_vcf} enables you to query variants within a VCF
}
\examples{
query_vcf("pot-2","II:1-10000","WBGene00010785")
}
