% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenotypes.R
\name{resolve_isotype}
\alias{resolve_isotype}
\title{\code{resolve_isotype} determines the isotype from a strain name.}
\usage{
resolve_isotype(strain_list)
}
\arguments{
\item{name_list}{A list of strains}
}
\value{
A list of isotypes.
}
\description{
This is the detail section if you want to fill out in the future
}
