% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variant_correlation.R
\name{snpeff}
\alias{snpeff}
\title{Browse Variant Info}
\usage{
snpeff(regions, severity = c("HIGH", "MODERATE"), long = TRUE,
  impute = TRUE)
}
\arguments{
\item{query}{A gene name, region, or wormbase identifier to query.}
}
\value{
Outputs a data frame that contains phenotype data, mapping data, and gene information for highly correlated variants in a particular QTL confidence interval.
}
\description{
\code{snpeff} enables you to query variants called and annotated by the \href{http://www.andersenlab.org}{Andersen Lab}.
}
\examples{
snpeff(c("pot-2","II:1-10000","WBGene00010785"))
}

