% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snps}
\alias{snps}
\title{snps}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38588 rows and 334 columns.}
\usage{
snps
}
\description{
snps
}
\keyword{datasets}
