% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spike.R
\name{spike}
\alias{spike}
\title{Spike in an effect for GWAS mapping}
\usage{
spike(meandiff, sd, snps, snpindex)
}
\arguments{
\item{sd}{The standard deviation for both populations}

\item{snps}{The snp set being used}

\item{snpindex}{The index of the snp upon which to spike the effect}

\item{meadiff}{The abolute difference in means for the two allele populations}
}
\value{
A phenotype data frame with the fake spiked data, ready for process_pheno
}
\description{
This function generates fake data for testing GWAS mappings, with the option
to generate QTL.
}

