% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{strain_isotype}
\alias{strain_isotype}
\title{Strain Isotype (Dataset)}
\format{tbl_df}
\usage{
strain_isotype
}
\description{
Dataset containing strain and isotype information
}
\keyword{datasets}

