% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{tajimas_d}
\alias{tajimas_d}
\title{GWAS Mappings}
\usage{
tajimas_d(vcf_path = get_vcf, chromosome = "II",
  interval_start = 11021073, interval_end = 12008179, window_size = 300,
  slide_distance = 100, samples = colnames(snps[, 3:ncol(snps)]),
  outgroup = "N2", site_of_interest = 11875145)
}
\arguments{
\item{vcf_path}{character value of the directory path to the VCF file.}

\item{chromosome}{character value corresponding to chromosome of interest. This should be the same chromosome name that is present in the VCF file}

\item{interval_start}{numeric value start of interval of interest}

\item{interval_end}{numeric value end of interval of interest}

\item{window_size}{size of window to calculate Tajima's D in SNPs (default = 300)}

\item{slide_distance}{number of SNPs to shift window (default = 100)}

\item{samples}{character vector of strains to use for analysis (default is 152 strain list present in cegwas package)}

\item{outgroup}{character value corresponding to strain (or strains) to serve as reference (default value is "N2")}

\item{site_of_interest}{numeric value, if specified adds a line to the plot at this position.}
}
\value{
Outputs a list, where the first element is a dataframe containing Tajima's D calculations and the second is a ggplot2 object visualizing Tajima's D across the interval
}
\description{
\code{tajimas_d} uses the PopGenome package to calculate Tajima's D across an interval.
}
\details{
This is the detail section if you want to fill out in the future
}

