% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_correlation.R
\name{variant_correlation}
\alias{variant_correlation}
\title{Process QTL Intervals}
\usage{
variant_correlation(df, quantile_cutoff_high = 0.9,
  quantile_cutoff_low = 0.1, variant_severity = c("MODERATE", "SEVERE"),
  gene_types = "ALL", kin = cegwas::kinship, condition_trait)
}
\arguments{
\item{df}{is a dataframe that is output from the \code{process_mappings} function}

\item{quantile_cutoff_high}{is a quantile cutoff that determines what variants to keep, default is to keep all variants with correlation coefficients greater than the 90th quantile}

\item{quantile_cutoff_low}{is a quantile cutoff that determines what variants to keep, default is to keep all variants with correlation coefficients less than the 10th quantile}

\item{variant_severity}{what variants to look at from snpeff output}

\item{gene_types}{what gene types to look at from snpeff output}

\item{kin}{is a strain by strain relatedness matrix you want to correct you trait with}
}
\value{
Outputs a list. Each list contains two data frames, the first contains mapping information (e.g. log10p, confidence interval start and stop), phenotype information, and gene ids. 
The second element of the list contains more detailed gene information
}
\description{
\code{variant_correlation} Returns all highly correlated variants in a QTL confidence interval
}
\details{
This function losely identifies unique intervals for QTL in a data set and browses a whole-genome variant set for variants that are highly correlated with the phenotype that gave rise to the QTL. Spearman rank correlation is used. 
Heterozygotes are ignored in the analysis. Correlations are only calculated for variants that are present in at least 5% of the assayed strains. Additionally, variant information needs to have been
acquired for at least 80% of the phenotyped strains, this removes the possibility of discrepency between correlated variants and the correlation that led to the QTL.
}

