% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcf.R
\name{vcf_to_matrix}
\alias{vcf_to_matrix}
\title{VCF To Matrix}
\usage{
vcf_to_matrix(vcf, allele_freq = 0, tag_snps = NA)
}
\arguments{
\item{vcf}{a bcf, vcf, or vcf.gz file}

\item{allele_freq}{allele frequency to filter on. Default is 0}

\item{variants}{A set of variants to filter on. By default, all variants are taken.}
}
\value{
Matrix of genotype calls
}
\description{
\code{vcf_to_matrix} converts variant calls from a bcf, vcf.gz or vcf file into an R dataframe.
Uses only biallelic variants. Heterozygous calls are ignored. Requires bcftools.
}
\seealso{
\link{generate_kinship}, \link{generate_mapping}
}

