% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{fetch_id_type}
\alias{fetch_id_type}
\title{Fetch Variant Type}
\usage{
fetch_id_type(id_type = NA)
}
\arguments{
\item{id_type}{Type of genetic element. Omit for list of options.}
}
\value{
Vector of identifiers.
}
\description{
\code{fetch_id_type} Fetches wormbase identifiers of a certain class. These can be piped into \code{snpeff}
}
\examples{
fetch_id_type("lincRNA")
}
