% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{generate_kinship}
\alias{generate_kinship}
\title{Generate Kinship}
\usage{
generate_kinship(vcf)
}
\arguments{
\item{a}{vcf file.}
}
\value{
A kinship matrix.
}
\description{
\code{generate_kinship} uses \code{\link{vcf_to_matrix}} and the \code{\link[rrBLUP]{A.mat}} function from \code{\link{rrBLUP}}.
This function is used to generate a kinship matrix for mapping in conjunction with \code{\link{gwas_mappings}}.
}
\seealso{
\link{vcf_to_matrix} \link{generate_kinship}
}

