% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{global_allele_distribution}
\alias{global_allele_distribution}
\title{Global distribution of allele}
\usage{
global_allele_distribution(location = "top-2", variant = 797,
  colors = c("purple", "salmon"))
}
\arguments{
\item{location}{A gene or position. Examples: \code{top-2}, \code{II:14425175}, or "WBGene00010785"}

\item{variant}{character value corresponding amino acid change you are interested in plotting. Currently only supports visualization of SNVs that alter an amino acid residue.}

\item{colors}{character vector containing two colors. The first element of the vector will be the ALT genotype and the second will be the REF genotype. Default values are c("purple","salmon").}
}
\value{
Outputs a ggplot object containing a visualization of the global distribution of allele of interest
}
\description{
\code{global_allele_distribution} uses the ggplot2 package to plot the global distribution of an allele of interest.
}
\details{
Specify either a \code{position} or a \code{gene} and amino acid \code{variant} position.
}

