% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_correlation.R
\name{snpeff}
\alias{snpeff}
\title{Browse Variant Info}
\usage{
snpeff(regions, severity = c("HIGH", "MODERATE"), long = TRUE,
  remote = FALSE)
}
\arguments{
\item{regions}{query A gene name, region, or wormbase identifier to query.}

\item{severity}{A vector with variants of given severities (LOW, MODERATE, HIGH, MODIFIER). Default takes moderate and high.}

\item{long}{Return dataset in long or wide format. Default is to return in long format.}

\item{remote}{Use remote data. Checks for local data if possible. False by default.}
}
\value{
Outputs a data frame that contains phenotype data, mapping data, and gene information for highly correlated variants in a particular QTL confidence interval.
}
\description{
\code{snpeff} enables you to query variants called and annotated by the \href{http://www.andersenlab.org}{Andersen Lab}.
}
\examples{
snpeff(c("pot-2","II:1-10000","WBGene00010785"))
}

