% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{snps}
\alias{snps}
\title{SNP set (Dataset)}
\format{A data frame with snps as rows and columns as strains:
\describe{
  \item{CHROM}{chromosome in roman numerals}
  \item{POS}{physical position in genome}
  \item{AB1}{first strain in collection, followed by all other strains}
  ...
}}
\description{
Genotype information for C. elegans wild isolates. SNPs used from RADseq data set but
includes all 124 wild isolates from whole-genome sequencing project
}

