% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike.R
\name{spike}
\alias{spike}
\title{Spike in an effect for GWAS mapping}
\usage{
spike(snps, snpindex)
}
\arguments{
\item{snps}{The snp set being used}

\item{snpindex}{The index of the snp upon which to spike the effect}
}
\value{
A phenotype data frame with the fake spiked data, ready for process_pheno
}
\description{
This function generates fake data for testing GWAS mappings, with the option
to generate QTL.
}
