% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{strain_isotype}
\alias{strain_isotype}
\title{Strain Isotypes (Dataset)}
\format{"strain" isotype" "latitude" "longitude""sequenced" "previous_names" "warning_message" "alternative_name"}
\usage{
strain_isotype
}
\description{
Strain Isotype
}
\keyword{datasets}

