% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{strain_isotype}
\alias{strain_isotype}
\title{Strain Isotype (Dataset)}
\format{tbl_df}
\description{
Table of strains and their respective isotypes in addition to strains with potential problems.
}

